Add umode basic infrastructure.
This patch adds the basic structure for umode. A umode task is created
in each physical CPU.

A CPU can run a umode task by activating first.

The activated state of a umode task is described by a type. When the
variable of that type is dropped, the umode task of the CPU is
re-initialized.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>