Implement umode requests in hypervisor using umode-api.
Introduce a request-based model for the interaction
between umode and the hypervisor.

The umode task for each CPU will be started at boot, execute the
initialization and wait for a request.

From then on the hypervisor can issue requests at the local CPU umode
task and get a result.

This patch introduce the hypervisor side of these changes.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>