Implement u-mode requests in hypervisor using u-mode-api.
Introduce a request-based model for the interaction
between u-mode and the hypervisor.

The u-mode task for each CPU will be started at boot, execute the
initialization and wait for a request.

From then on the hypervisor can issue requests at the local CPU u-mode
task and get a result.

This patch introduce the hypervisor side of these changes.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>