HypMap: Add ability to map guest pages in U-mode.
To avoid the need for a VA-allocator of some type, assign a part of
the U-mode VA space for a number (currently 2) of equal sized 'slots'.

The user will be reponsible to allocate a slot to each input/output
area.

Final value of number of slots and the size of each is not fixed, and
can be changed easily for future needs.

Also add a UmodeMapping type, that represents the mapping of a U-mode
slot. On drop, the mapping will be unmapped and the tlb flushed.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>