HypMap: Add low-level functions to map guest pages in U-mode.
To avoid the need for a VA-allocator, assign a part of the U-mode VA
space for a number (currently 2) of equal sized 'slots'.

The caller will be responsible to allocate a slot to each input/output
area.

Final value of number of slots and the size of each is not fixed, and
can be easily changed in the future.

Also add UmodeSlotMapper, a mapper type for mapping guest pages into
U-mode slots.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>