Separate between U-mode requests errors from U-mode execution errors.
In case U-mode would send a wrong hypcall, we would previously simply
return a U-mode error. This is wrong. Fix it (and clean up) by
distinguishing APIs error coming from the hypervisor (that are to be
considered fatal U-mode execution errors) and API errors that result
from a completed operation in U-mode.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>