Salus: Switch stack to VA mapped.
This patch allows both bootstrap CPU and secondary CPUs to switch to
VA mapped stack once the pagetable is set (before calling `launch`).

Size of stack for secondary CPUs is increased to the same size of boot
cpu (defined by the linker). We already were running out of stack with
8 pages, but it was difficult to catch on a physically mapped stack.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>